refactor: rename local agents module
