+++ steps/step5_entity_instances.py
@@ -8,7 +8,7 @@ from pydantic import ValidationError
 
 from agents import RunConfig, RunResult, TResponseInputItem  # type: ignore[attr-defined]
 
-from ..agents import entity_instance_extractor_agent
+from ..workflow_agents import entity_instance_extractor_agent
 from ..config import (
     ENTITY_INSTANCE_MODEL,
     ENTITY_INSTANCE_OUTPUT_DIR,
