+++ steps/step6a_relationship_types.py
@@ -9,7 +9,7 @@ from pydantic import ValidationError
 
 from agents import RunConfig, RunResult, TResponseInputItem  # type: ignore[attr-defined]
 
-from ..agents import relationship_type_identifier_agent
+from ..workflow_agents import relationship_type_identifier_agent
 from ..config import (
     RELATIONSHIP_MODEL,
     RELATIONSHIP_OUTPUT_DIR,
